Add scheduler to update sample time series.
Ofelia is used as a scheduler to run bash scripts on the timescaledb container that update the database with new values. The schedules that are run are controlled by the config.ini file, and files are staged in the container using docker-compose.yml. A new Dockerfile was added to build the timescaledb image so that the xmlstarlet and curl packages can be added to that container.